backend/requirements.txt
fastapi
uvicorn
ollama
python-dotenv
sqlalchemy
pydantic
faiss-cpu
sentence-transformers
PyPDF2
python-docx
python-multipart
numpy
Pillow
opencv-python
pytesseract
